requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
jinja2==3.1.4
python-multipart==0.0.9
websockets==12.0
httpx==0.27.0
google-generativeai==0.5.4
openai==1.30.1
aiofiles==23.2.1
